Umstellung der Distanz auf Long

+++ JavaMaps/src/de/javamaps/items/Vertex.java
@@ -9,7 +9,7 @@ public class Vertex {
 	private final String name;
 	private Vertex previous = null;	
 	private int dist = Integer.MAX_VALUE; /// ersatz für unendlich //bisheriger Weg zu diesem Knoten
-	private Map<Vertex, Integer> adjacentVertex = new HashMap<>();
+	private Map<Vertex, Long> adjacentVertex = new HashMap<>();
 
 	public Vertex(String name) {
 		this.name = name;
@@ -35,11 +35,11 @@ public class Vertex {
 		this.dist = dist;
 	}
 
-	public Map<Vertex, Integer> getAdjacentVertex() {
+	public Map<Vertex, Long> getAdjacentVertex() {
 		return adjacentVertex;
 	}
 
-	public void setAdjacentVertex(Map<Vertex, Integer> adjacentVertex) {
+	public void setAdjacentVertex(Map<Vertex, Long> adjacentVertex) {
 		this.adjacentVertex = adjacentVertex;
 	}
 }
